requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
charset-normalizer==3.2.0
colorama==0.4.6
discord.py==2.3.2
frozenlist==1.4.0
idna==3.4
iniconfig==2.0.0
multidict==6.0.4
packaging==23.1
pluggy==1.3.0
pytest==7.4.2
pytest-asyncio==0.21.1
python-dotenv==1.0.0
yarl==1.9.2
